perf: Treat the kernel like [kernel.kallsyms]
The perf ABI knows about the kernel, and there are bits in the data format
for kernel mmaps and traces.  When we run Linux's perf report, we can point
it at our symbol table and it can resolve the symbols.

Previously, we were treating the kernel like it was a user binary and had
to hack up Linux's perf to handle it accordingly.  Now we're just emitting
the right thing.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>