Remove kernel path and hostname from #version
Perf had been using the kernel path so that it could emit it in
perf.data.  We no longer need that.  The kernel path was the path on the
machine on which the kernel was built, which assumes the perf report would
be analyzed on that same machine.

Likewise, this removes the hostname from #version.  This hostname wasn't
the name of the Akaros machine; it was the name of the machine on which
Akaros was built.  I don't see a good reason to keep this around.

If we need either of these in the future, we can probably replace it with
build-id (from the linker).

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>